Notification/requirements.txt
azure-functions
requests
python-dotenv
